Task 0

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -15,7 +15,7 @@ int sum_them_all(const unsigned int n, ...)
 
 	va_start(ptr, n);
 
-	for (i = 0; i < n; i++)
+	for (i = 0; (unsigned int) i < n; i++)
 		sum += va_arg(ptr, int);
 
 	va_end(ptr);
